+++ Assets/Scripts/UI/TakeDamageEvent.cs
@@ -4,7 +4,7 @@ using UnityEngine.Events;
 namespace RPG.UI
 {
     [System.Serializable]
-    public class TakeDamageEvent : UnityEvent<int, DamageType, bool, WeaponConfig>
+    public class TakeDamageEvent : UnityEvent<float, DamageType, bool, WeaponConfig>
     {
         
     }
